Shove existing bits into new form around kputil
A lot of moving around and stuff around a new build target named kputil which
will have all of the functionality I plan to implement as modes, rather than
making lots of small separate programs.  Move a lot of stuff from merge.c to
util.c and make a list.c for listing the contents of a database (currently
ignores groups...).  Still working on a good way to display a database in text.