Add brief output type to list
The masking feature in list is perhaps the most obtuse way to expose the
ability to show or hide fields but it is also the easiest to implement.  To
make it easier to deal with, this adds a brief mode that sets the masks to
UUID/username/title and Group ID/name.